Update WhatsApp_test.ino
This is a later version than I tested. There was an issue with char array lengths which should be fixed now and the cool room works to <5C.  At the shed it would initially connect but then nothing after 10 mins so I assume its being kicked off the wifi. need to rework this so that it reconnects before sending a message.